pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.27.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "sift-stack-py>=0.7.0",
    "tensorflow>=2.14.0",
    "werkzeug>=3.1.3",
    "pillow>=11.3.0",
    "requests>=2.32.4",
    "numpy>=2.3.2",
]
